Results RSD src + bin

+++ benchmarks/RSD/compile_and_rename_all.sh
@@ -2,6 +2,11 @@
 
 BASE_DIR="${1:-.}"
 
+if [ -z "$BASE_DIR" ]; then
+    echo "Errore: specify a path."
+    exit 1
+fi
+
 rm -rf bins
 
 find "$BASE_DIR" -type f -name "*.sol" | while read -r file; do
